Correct HTTP responses
Need to make response show on screen
