+++ androidapp/Connexus/app/src/main/java/com/green/apt/connexus/ViewAllActivity.java
@@ -2,13 +2,27 @@ package com.green.apt.connexus;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.GridView;
+import android.widget.TextView;
 import android.widget.Toast;
 import android.widget.AdapterView.OnItemClickListener;
 
+import com.android.volley.Request;
+import com.android.volley.RequestQueue;
+import com.android.volley.Response;
+import com.android.volley.VolleyError;
+import com.android.volley.toolbox.StringRequest;
+import com.android.volley.toolbox.Volley;
+
 public class ViewAllActivity extends AppCompatActivity {
+
+    private static final String BASE_URL = "https://apt-miniproject-greenteam-v2.appspot.com/api/";
+    private static String relativeUrl = "view_all";
+    private static String streams = "All Streams";
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -24,5 +38,41 @@ public class ViewAllActivity extends AppCompatActivity {
                         Toast.LENGTH_SHORT).show();
             }
         });
+
+        this.getAllStreams();
+
+        TextView mTextView = (TextView) findViewById(R.id.textView);
+        mTextView.setText(streams);
+
+    }
+
+    private void getAllStreams(){
+        // Instantiate the RequestQueue.
+        RequestQueue queue = Volley.newRequestQueue(this);
+        String url =getAbsoluteUrl(relativeUrl);
+        Log.d("ViewAllActivity","URL: "+ url);
+
+        // Request a string response from the provided URL.
+        StringRequest stringRequest = new StringRequest(Request.Method.GET, url,
+                new Response.Listener<String>() {
+                    @Override
+                    public void onResponse(String response) {
+                        // Display the first 500 characters of the response string.
+                        Log.d("ViewAllActivity","response: "+ response.substring(0,500));
+                        streams = "Response is: "+ response.substring(0,500);
+                    }
+                }, new Response.ErrorListener() {
+            @Override
+            public void onErrorResponse(VolleyError error) {
+                Log.d("ViewAllActivity","response: "+ "that didn't work");
+                streams = "That didn't work!";
+            }
+        });
+        // Add the request to the RequestQueue.
+        queue.add(stringRequest);
+    }
+
+    private static String getAbsoluteUrl(String relativeUrl) {
+        return BASE_URL + relativeUrl;
     }
 }
